# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.CaasWebhookPayload:
    properties:
      data:
        properties:
          amount:
            type: string
          blind_index:
            type: string
          currency:
            type: string
          phone:
            type: string
          reference:
            type: string
          status:
            type: string
          tx_hash:
            type: string
        type: object
      event:
        type: string
    type: object
  handlers.CreateCardRequest:
    properties:
      card_art_id:
        example: art_abstract_01
        type: string
      color_theme:
        example: midnight_blue
        type: string
      funding_account_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      funding_wallet_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      name:
        example: My Daily Card
        type: string
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        properties:
          code:
            example: VALIDATION_ERROR
            type: string
          message:
            example: phone is required
            type: string
        type: object
      success:
        example: false
        type: boolean
    type: object
...
  handlers.MessageResponse:
    properties:
      message:
        example: OTP sent
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  handlers.UpdateCardRequest:
    properties:
      color_theme:
        example: sunset_orange
        type: string
      name:
        example: Travel Card
        type: string
      status:
        example: frozen
        type: string
    type: object
...
  handlers.VirtualCardData:
    properties:
      card_art_id:
        type: string
      color_theme:
        type: string
      created_at:
        type: string
      expiry:
        type: string
      id:
        type: string
      masked_pan:
        type: string
      name:
        type: string
      status:
        type: string
    type: object
  handlers.VirtualCardListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/handlers.VirtualCardData'
        type: array
      success:
        example: true
        type: boolean
    type: object
  handlers.VirtualCardResponse:
    properties:
      data:
        $ref: '#/definitions/handlers.VirtualCardData'
      success:
        example: true
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: engineering@rach.finance
    name: Rach Finance Engineering
  description: |-
    DigitalFX is a multi-currency fintech platform for Central/West Africa.
    It combines fiat accounts (XAF, XOF, USD, GBP, EUR), HD crypto wallets (BIP-44 via Rach WaaS),
    and stablecoin P2P transfers (USDC/USDT via Rach CaaS ERC-4337 SCWs).
    Mobile Money deposits and withdrawals are handled via HUB2 (MTN, Orange).
  title: DigitalFX API
  version: "1.0"
paths:
...
  /cards:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.VirtualCardListResponse'
      security:
      - BearerAuth: []
      summary: List virtual cards
      tags:
      - Cards
    post:
      consumes:
      - application/json
      parameters:
      - description: Card creation details
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateCardRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.VirtualCardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create virtual card
      tags:
      - Cards
  /cards/{id}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Card ID
        in: path
        name: id
        required: true
        type: string
      - description: Update payload
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateCardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.VirtualCardResponse'
      security:
      - BearerAuth: []
      summary: Update virtual card
      tags:
      - Cards
...
  /webhooks/caas:
    post:
      consumes:
      - application/json
      description: Receiver for Rach CaaS account/settlement events (funding, on-chain
        credit, P2P transfer, off-ramp). Follows the standard Rach webhook signing
        — header `X-Webhook-Signature` = hex HMAC-SHA256 of the raw request body,
        keyed with the CaaS webhook secret (CAAS_WEBHOOK_SECRET), which is SEPARATE
        from the WaaS secret. Verification also tolerates alternate header names and
        base64 as a hardening fallback. On a credit/deposit event the owning user
        (resolved by phone) is notified, always labeled iUSD (never USDC). Always
        returns 200 on business-logic issues so the provider does not retry.
      parameters:
      - description: hex HMAC-SHA256(rawBody, CaaS secret)
        in: header
        name: X-Webhook-Signature
        required: true
        type: string
      - description: Webhook payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CaasWebhookPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Rach CaaS Webhook (Instant USD / iUSD)
      tags:
      - webhooks
...
securityDefinitions:
  BearerAuth:
    description: 'JWT access token — prefix with "Bearer ": Bearer eyJhbGci...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"